requirements.txt
fastapi
uvicorn
transformers
torch
Pillow
python-multipart
faiss-cpu
